% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpathwaysdocs.R
\docType{package}
\name{rpathwaysdocs}
\alias{rpathwaysdocs}
\alias{rpathwaysdocs-package}
\title{rpathwaysdocs}
\description{
Results pathways are usually based on a theory of change describing
how change is assumed to come about through intervention/s in a prevailing
situation. The theory is laid out in a diagram showing the connections
between interventions and outcomes - the results pathways. This package
contains text datasets of various documents describing current thinking
on the use of results pathways for monitoring and evaluations various
projects.
}
