% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpathwaysdocs.R
\docType{data}
\name{spring_brief_1}
\alias{spring_brief_1}
\title{spring_brief_1}
\format{A dataframe with 396 rows and 2 columns:
\describe{
\item{\code{linenumber}}{Line number}
\item{\code{text}}{Character string of about 70 characters each}
}}
\source{
\url{https://www.spring-nutrition.org/sites/default/files/publications/briefs/spring_understandingpathways_brief_1.pdf}
}
\usage{
spring_brief_1
}
\description{
Understanding and Applying Primary Pathways and Principles
}
\details{
A dataset containing the text of Understanding and Applying Primary Pathways
and Principles. The UTF-8 plain text is divided into elements of
up to about 70 characters each. Some elements are blank.
}
\keyword{datasets}
