% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpathwaysdocs.R
\docType{data}
\name{spring_brief_4}
\alias{spring_brief_4}
\title{spring_brief_4}
\format{A dataframe with 208 rows and 2 columns:
\describe{
\item{\code{linenumber}}{Line number}
\item{\code{text}}{Character string of about 70 characters each}
}}
\source{
\url{https://www.spring-nutrition.org/sites/default/files/publications/briefs/spring_womensempowerment_brief_4_0.pdf}
}
\usage{
spring_brief_4
}
\description{
Understanding the Women’s Empowerment Pathway
}
\details{
A dataset containing the text of Understanding the Women’s Empowerment Pathway.
The UTF-8 plain text is divided into elements of up to about 70 characters
each. Some elements are blank.
}
\keyword{datasets}
