% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpathwaysdocs.R
\docType{data}
\name{vogel_toc_review_2012}
\alias{vogel_toc_review_2012}
\title{vogel_toc_review_2012}
\format{A dataframe with 2152 rows and 2 columns:
\describe{
\item{\code{linenumber}}{Line number}
\item{\code{text}}{Character string of about 70 characters each}
}}
\source{
\url{http://www.theoryofchange.org/pdf/DFID_ToC_Review_VogelV7.pdf}
}
\usage{
vogel_toc_review_2012
}
\description{
Review of the use of Theory of Change in international development
}
\details{
A dataset containing the text of Review of the use of Theory of Change in
international development. The UTF-8 plain text is divided into elements of
up to about 70 characters each. Some elements are blank.
}
\keyword{datasets}
