% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpathwaysdocs.R
\docType{data}
\name{world_bank_results_framework_2012}
\alias{world_bank_results_framework_2012}
\title{world_bank_results_framework_2012}
\format{A dataframe with 1352 rows and 2 columns:
\describe{
\item{\code{linenumber}}{Line number}
\item{\code{text}}{Character string of about 70 characters each}
}}
\source{
\url{www.worldbank.org}
}
\usage{
world_bank_results_framework_2012
}
\description{
Papua New Guinea's Alotau Accord for a Government of Reconciliation, Unity and Hope
}
\details{
A dataset containing the text of Papua New Guinea's Alotau Accord for a
Government of Reconciliation, Unity and Hope. The UTF-8 plain text is
divided into elements of up to about 70 characters each. Some elements are
blank.
}
\keyword{datasets}
